@@ -12,7 +12,7 @@ module OmniauthRegistrationsControllerExtend
 
     after_action :grant_omniauth_authorization, except: [:logout]
 
-    skip_before_action :verify_authenticity_token, if: :is_saml_callback?
+    skip_before_action :verify_authenticity_token, if: :saml_callback?
 
     def new
       @form = form(Decidim::OmniauthRegistrationForm).from_params(user_params)
@@ -200,7 +200,7 @@ module OmniauthRegistrationsControllerExtend
       current_organization.enabled_omniauth_providers[provider.to_sym][:provider_name].presence || provider.capitalize
     end
 
-    def is_saml_callback?
+    def saml_callback?
       request.path.end_with?("saml/callback") || request.path.end_with?("csam/callback")
     end
   end
