@@ -1,33 +1,33 @@
 # frozen_string_literal: true
 
-require 'omniauth-saml'
-require 'savon'
-require 'signer'
-require 'nokogiri'
+require "omniauth-saml"
+require "savon"
+require "signer"
+require "nokogiri"
 
 module OmniAuth
   module Strategies
     class EidSaml < OmniAuth::Strategies::SAML
       option :name, :saml
 
-      option :origin_param, 'redirect_url'
+      option :origin_param, "redirect_url"
 
-      option :authn_context_comparison, 'minimum'
-      option :name_identifier_format, 'urn:oasis:names:tc:SAML:2.0:nameid-format:transient'
-      option :uid_attribute, 'urn:be:fedict:iam:attr:fedid'
-      option :attribute_service_name, 'Eidas extra attributes'
+      option :authn_context_comparison, "minimum"
+      option :name_identifier_format, "urn:oasis:names:tc:SAML:2.0:nameid-format:transient"
+      option :uid_attribute, "urn:be:fedict:iam:attr:fedid"
+      option :attribute_service_name, "Eidas extra attributes"
       option :attribute_statements,
-             name: %w[uid name],
-             email: %w[mail email],
-             first_name: %w[surname first_name firstname firstName],
-             last_name: %w[givenName last_name lastname lastName],
-             # default_locale: ['locale', 'urn:be:fedict:iam:attr:locale'],
-             locale: ['PrefLanguage', 'pref_language', 'preflanguage', 'locale', 'urn:be:fedict:iam:attr:locale'],
-             nickname: ['uid'],
-             rrn: %w[egovNRN egovnrn egov_nrn nrn rrn]
-      # authentication_method: ['authenticationmethod'],
-      # authentication_level: ['urn:be:fedict:iam:attr:authenticationmethod'],
-      # authentication_context: ['urn:be:fedict:iam:attr:context']
+             name: %w(uid name),
+             email: %w(mail email),
+             first_name: %w(surname first_name firstname firstName),
+             last_name: %w(givenName last_name lastname lastName),
+             # default_locale: ("locale", "urn:be:fedict:iam:attr:locale"),
+             locale: ["PrefLanguage", "pref_language", "preflanguage", "locale", "urn:be:fedict:iam:attr:locale"],
+             nickname: ["uid"],
+             rrn: %w(egovNRN egovnrn egov_nrn nrn rrn)
+      # authentication_method: ["authenticationmethod"],
+      # authentication_level: ["urn:be:fedict:iam:attr:authenticationmethod"],
+      # authentication_context: ["urn:be:fedict:iam:attr:context"]
       option :idp_cert_fingerprint_validator, ->(fingerprint) { fingerprint }
       option :force_authn, true
       option :security,
@@ -71,4 +71,4 @@ module OmniAuth
   end
 end
 
-OmniAuth.config.add_camelization 'saml', 'SAML'
+OmniAuth.config.add_camelization "saml", "SAML"
