@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
-require 'decidim/core/test/factories'
-require 'decidim/verifications/omniauth/test/factories'
+require "decidim/core/test/factories"
+require "decidim/verifications/omniauth/test/factories"
